+++ sql-planner/src/ir.rs
@@ -646,12 +646,12 @@ bitflags! {
 
 impl ExplainOptions {
     #[inline(always)]
-    pub fn facettes() -> Self {
+    pub fn facets() -> Self {
         Self::Logical | Self::Raw
     }
 
-    pub fn has_facette(&self) -> bool {
-        self.intersects(Self::facettes())
+    pub fn has_facet(&self) -> bool {
+        self.intersects(Self::facets())
     }
 }